[refactor] リファクタリング
